requirements.txt
socket
pybluez
gitpython
